Fix assert() side effects
Squash of several feature/assert-side-effects WIP commits

Fixes the use of function calls with side-effects within assert(), assignment within assert(),
and the asserts in the type_malloc macros

This is only intended to transition the binaries created by a "release" build from broken to working. I intend to revist
this to make the code less ugly (biggest issue is unused var warnings when asserts are disabled).


Former-commit-id: 4853a22d82cc7fee968131a59d9f9adbf8b8bb39